Update installation.rst
Hey, it was my first time using saas and it's a bit confusing if you want to change the css. I hope these lines are clear and will help someone without any knowledge in saas from now on. =) Feel free to improve them. I feel like the "installation" section may not be the right place for that.